decompose TransactionDataUtils
