+++ waltid-libraries/protocols/waltid-openid4vp-wallet/src/commonMain/kotlin/id/waltid/openid4vp/wallet/presentation/W3CPresenter.kt
@@ -6,8 +6,8 @@ import id.walt.credentials.signatures.sdjwt.SelectivelyDisclosableVerifiableCred
 import id.walt.crypto.keys.Key
 import id.walt.dcql.DcqlDisclosure
 import id.walt.dcql.DcqlMatcher
-import id.walt.verifier.openid.TransactionDataUtils
 import id.walt.verifier.openid.models.authorization.AuthorizationRequest
+import id.walt.verifier.openid.transactiondata.filterTransactionDataForCredentialId
 import id.walt.w3c.PresentationBuilder
 import id.waltid.openid4vp.wallet.WalletPresentFunctionality2.createKeyBindingJwt
 import io.github.oshai.kotlinlogging.KotlinLogging
@@ -47,7 +47,7 @@ object W3CPresenter {
                 audience = authorizationRequest.clientId,
                 selectedDisclosures = disclosuresToPresent, // Pass the actual disclosures for sd_hash
                 holderKey = holderKey,
-                transactionData = TransactionDataUtils.filterTransactionDataForCredentialId(
+                transactionData = filterTransactionDataForCredentialId(
                     transactionData = authorizationRequest.transactionData,
                     credentialId = matchResult.originalQuery.id,
                 ),
