+++ waltid-libraries/protocols/waltid-openid4vp/src/commonMain/kotlin/id/walt/verifier/openid/transactiondata/TransactionDataConstants.kt
@@ -1,0 +1,7 @@
+package id.walt.verifier.openid.transactiondata
+
+const val DEFAULT_HASH_ALGORITHM = "sha-256"
+const val MDOC_DEVICE_SIGNED_NAMESPACE = "org.waltid.openid4vp.transaction_data"
+const val DEMO_TRANSACTION_DATA_TYPE = "org.waltid.transaction-data.payment-authorization"
+
+val SUPPORTED_TRANSACTION_DATA_TYPES = setOf(DEMO_TRANSACTION_DATA_TYPE)
