+++ waltid-libraries/protocols/waltid-openid4vp/src/commonMain/kotlin/id/walt/verifier/openid/transactiondata/TransactionDataSelection.kt
@@ -1,0 +1,8 @@
+package id.walt.verifier.openid.transactiondata
+
+fun filterTransactionDataForCredentialId(
+    transactionData: List<String>?,
+    credentialId: String,
+): List<String> = decodeList(transactionData.orEmpty())
+    .filter { credentialId in it.transactionData.credentialIds }
+    .map { it.encoded }
